Parser Rescaffolding
This commit rescaffolds the Parser. At the moment, it's heavily modeled after
the Pratt Parser found in "Crafting Interpreters". Because the Scanner today
returns the entire lexed corpus, we receive and take ownership of an iterator
containing the tokens. This is up for refactoring later.

This commit is a squashed commit. I don't _love_ that I did this, but there
were more divergences/wip commits/wtf moments that it honestly made it easier
in the `main` branch just to say "here it is".

Changes Include:
- Rework the scanner to receive `String` as input as an argument to the ctor
itself. This may not be a long term solution (size is a concern)
- Rework some ownership issues in main.rs
- Implements the whole scaffolded Pratt Parser
   - Add Precedence enum
   - Add methods for number, binary, unary and grouping operations
   - Add mechanism to iterate over iterator of tokens
   - Add parse rule look up table (needs refactoring)
   - Stub out error handling
   - Stub out initial data emitting