feature(repl): run repl more than once per session
allow the repl to evaluate more than one line of code. this commit moves
core STDIN functionality into `lib.rs` while simultanesouly removing the
'stop and rerun/re-init' the VM functionality that was there previously.

this commit introduces a flag, '-k' to allow for backwards
compatiabilty. this allows the brittle test suite to continue to run as
it did previously with no major changes (just yet).

this change revealed a small bug where the instruction pointer on the vm
was not being reset between invocations (because it didn't need to at
the time). this caused the chunk assigned to the VM and the way it was
indexed to fall out of sync. this has been fixed by reseting the
pointer.