feat(loops) add support for continue in while loops
this commit adds support for continue statements in while loops (as
that's the only type of loop we have at the moment). scope is tracked on
the parser level as a quasi-global variable. as while loops nest, we
keep track of the previous loop's starting value and restore it once
we're done parsing the loop