feat(conditional): implement conditional expressions
this commit adds support for conditional expressions (I.E. ternary
expressions), which take the form

> expr1 ? expr2 : expr3

a new parsing method, conditional_expression, has been added to the
parser module. this method handles both the question mark and colon
tokens, rather than trying to separate them into two separate fns. the
reason for which is that it _feels_ easier to keep track of a single
conditional expression this way, allowing the recusive nature of the
pratt parser to handle nested one (where for example, a whole
conditional expression is found in expr2 or expr3)

tests have been added to control_flow.rs and separate awk snippets
representing those tests have been added to the awk_examples directory,
in hopes that one dya we're able to automate the creation of them