refactor(vm): vm refactorings
refactor returning an OK() variant to within the matching arm for
OpCode::OpReturn. this allows us to remove the mut variable used for
tracking, and remove if statement at the beginning of the vm's loop

Use existing Result for returning an error when we attempt to print a
value when the stack is empty, rather than just calling panic!

Remove extraneous old_ip variable, which can be replaced with the VM#ip
field. This necessitates moving the increment of self.ip to the line
following the retrieval of the instruction