feat(fields): implement field variables
this commit introduces preliminary support for field variables
(variables prefixed with a 'sigil' - e.g. $2). field variables are
created by reading in data from a user/file, where preliminary support
for the former is added in this commit as well. specifically this
commit:

reworks the repl for the purposes of taking data in rather than
successive lines of awk code. this has multiple ramifications for
existing code and tests:

1. a temporary `-q` argument has been added to allow for running awk
   without any data input. this is temporary until `BEGIN` is
   implemented. existing e2e util crate helper fn now use this flag
2. an additional e2e util crate helper function has been created for
   inserting a single line of data into tests, `assert_input_with_data`
3. awk programs must be passed in at invocation time, and only one may
   be run per process `cargo run -- 'print $2;'`. single quotes should
   be used (for now) to avoid shell interpolation. if no program is
   provided, default to `print $0;`

stores the values read from the provided data in field variables.
the data provided by the user is split per the file separator
(whitespace only, see TODO below) and assigns the values to
numeric-based/index-based variables. for now, all field variables are
pushed as strings, but this will change (see bugs section below) to use
strnum.
- field variables are stored internally in an array to make the initial
  lookup aspect of the implementation easier. this may change in the
  future.
- the original string is stored at index 0. the reason for which being
  that when separating by whitespace, rebuilding that string can be a
  pain to do, especially if the WS is variable. again, this may change
  as needed.
---
this also opens up many new possibilities and items 'TODO':

1. implement `BEGINS`
2. implement file separators
3. implement reading a file from disk and running rawk against the data
   read

additional bugs were found during the implementation of this feature,
and will be addressed separately:

1. strnum does not work as expected, nor is it accounted for in the
   codebase/with field references
2. concatenation does not work for groupings, potentially other field
   (scope unknown)