feat(field_sep): initial field separator impl
this commit introduces the initial implementation of field separators,
provided by the user via the command line. this commit is limited
strictly to providing field separators in such a way, and does not
introduce reassigning the internal FS variable, allowing access for FS,
or any other related functionality.

several flags such as `is_quick` and `is_eval` are pushed into a new
struct, `RuntimeConfig`. this does not affect their current
ephemerability(?), but does provide a slightly cleaner contract between
the main and lib crates.

reading user data and splitting said data have been pulled out into two
functions. primarily because there is a high chance that the latter will
be used when other FS manipulations are introduced, but also the
function in which they were originally housed was starting to grow
larger than I'd have liked (purely subjective).

much of the splitting data logic is currently centered around getting
the rust impl of `split()` on string to adapt to the awk way of doing
things. i've temporarily decided not to try to write one myself, as that
feels like I'd be writing/copying the current split logic in the std lib
plus my own additions. if we need to cross that bridge in the future, so
be it.

all keys that are used by the CLI were extracted to const variables in
this commit. i started for FS and couldn't help myself. scope creep is
real.