refactor(prompt): consolidate arguments
consolidate arguments for the run_prompt() fn by removing the &[string]
argument. this arg was a holdover from the initial implementation of
field separators, and can now be removed. the name of the file (impl
TBD) recevied from the command line may now be used in its stead.

the filename received from the command line had been absorbed into the
runtime config