refactor(runtime): scan once per run
move scanning/tokeninzing to occur only once per invocation of r-awk,
rather than on every new invocation of `vm#interpret`. this was an
artifact of earlier versions of r-awk, where the repl would print the
result of every expression (which could be given to the repl instead of
data). this 'should' be faster/avoid unnecessary memory allocations, but
i did not measure here. the goal was really to have parsing failures
'fail fast' instead of requiring a line of user input to be read first,
to match awk implementations like gawk.

the parser could also be theoretically moved out in the same way. doing
so was avoided in this commit until i have a better idea of how i want
to handle multiple `Chunk` entities, which are owned by the `vm` and are
passed to the parser instance.