feat(action): require actions be wrapped
require actions be wrapped in curly braces as the first step in
supporting action+pattern code and multiple instances of the like.
update tests to be wrapped in braces.

small amounts of refactoring/reordering the awk examples was completed
while checking that all `awk_examples/` files were also wrapped in
braces