feat(pattern): initial pattern support
this commit introduces initial support for patterns in r-awk. patterns
are no longer implicitly 'true' in r-awk (unless omitted from the
program), and are properly parsed.

support for BEGIN and END will occur in a separate commit.

action parsing has been updated to be able to jump over an action in the
event the result of the pattern returns false, similar to how r-awk
parses if statements.

actions are now implicitly `print $0` in the event they are omitted