fix(comma): only perform comma concat in print stmt
this commit moves comma concatenation from being a general puprpose
operation to one that (for now) can only occur when printing an expr
list. this is a hold over from the days of being 'calculator' style
repl.

as a result, expressions occurring outside of a print statement like
`4,3` shall no longer evaluate to '4 3' (which is the correct behavior)