refactor(tests): create uniform testing method
this commit is the first step towards a better e2e testing strategy(TM).
remove all but one testing function from the tests/utils crate, and
unify all tests to use that function.

in the future, consider a builder to cut down on some of the noise/need
to have a 4-arg helper function