feat(awk): add program file support
this commit adds support for reading one or more files containing an
awk program whose path(s) are specified on the command line using the
`-f` flag, followed by the path to the file.

this commit also attempts to create a clearer distinction between awk
program files read from the command line, and data files to be used
against an awk program. full support for the latter is not implemented
here. however, the `runtime_config#file_name` field has been renamed to
`data_file_path`, and `None` is explicitly passed as the data argument
to creating a `RuntimeConfig` for the time being.

in the future, error messages should be more descriptive (i.e. when a
file does not exist, the name of the file should be reported back to the
user, rather than 'FileDoesNotExist'). reworking error handling is being
pushed back to a later effort.

a bug was discovered in the recently refactored utils testing package,
where expecting a run of r-awk to panic would fail. a workaround (with
TODOs) has been put in place, but will require additional work to make
the API easier to work with