chore(tests): apply must_use attr to CodeRunner
add the `#[must_use]` macro to `CodeRunner` to better protect against
cases where the `assert()` function is not called. this is accomplished
by passing ownership of an instance of `CodeRunner` through its
builders, rather than returning a mutable pointer to them.

as a result of this patch, the following code will now fail a lint as
intended:

```rust
    #[test]
    fn it_supports_an_awk_pattern_with_no_action() {
        utils::CodeRunner::init()
            .program("1 > 0")
            .stdin_data("Hello World")
            .expect_output("Hello World");
            // note the semicolon added above and assert() commented out
            // .assert()
    }
```

reminding us that the value `CodeRunner` hasn't been consumed (by
assert())