refactor(tests): improve readability with raw str
this commit uses rust's raw strings to improve the readability of some
tests found in the `tests/` dir. since the `CodeRunner` object takes a
string slice as an argument to its `program` builder, using raw strings
allows the argument to be passed to `program` without escaping quotes
(and is the main argument for readability here).

in some cases, new lines were added as well. these were subjective
descisions rather than ones tailored to a specific heuristic. (not my
favorite, but that's the "reason to the rhyme")