chore(deps): upgrade to latest clap v3 version (#12)
this commit bumps the version of clap to the last version published
under the v3 major-version banner. this is done to act as a 'quick
check' for any warnings/deprecations that _could_ have appeared during
the upgrade process from v3 to v4 (as we upgraded from an older v3
version). no warnings/deprecations were found