fix(field_variables): handle out of bounds unary ops (#13)
* fix(field_variables): handle out of bounds

this commit fixes a bug where an awk program that attempts to access
field variables during a unary operation, where the field variable did
not exist, causing a panic.

example:
```bash
$ rawk '{print +$3;}'
$ arg-1 arg-2
```

the root cause was a `UnaryPlus` || `UnaryMinus` operation would be
pushed onto the stack. when the vm attempted to perform the operation,
an empty string that was resolved from the field variable would be
detected. for this particular case when performing unary operations, no
handler for either unary op with a string (of any type) existed.

to resolve, we simply push a zero onto the stack to align with `awk`.