chore(deps): upgrade clap to v4 (#14)
this commit pulls the generation of a clap `Command` entity to a
separate function, in an effort to test the command being generated as a
part of the v4 upgrade process, as outlined in
https://github.com/clap-rs/clap/releases/tag/v4.0.0-rc.1. doing so
allows us to use the `debug_assert()` function provided by clap in a new
unit test to help programmatically verify (some) of the cli's
configuration.

the various `*_KEY` str values were hoisted to the outermost scope of
the `main` entrypoint to allow for reuse. they were not refactored under
a single namespace (e.g. an `enum`) to keep the commit smaller and more
focused. this allows us to use the str's in multiple places throughout
the file.

the generation of `version` using the clap builder api has been removed,
as it is no longer needed (and in fact errors in at runtime should we
add it as-is without additional modifications to the `Command` entity).

in order to differentiate as to whether the `-q` (quick) and `-k` (eval)
flags have been set by the user or by clap, we can no longer check if
either key is present (as it will always be present). we can however,
determine if the user provided it using it's origin (the command line vs
a default field). the generation of a `RuntimeConfig` has been updated
accordingly to make thsi distinction.

any remaining changes not expressly called out were done in accordance
to the migration guide linked earlier in this commit msg