feat(data): accept data files from command line (#17)
this commit introduces the first pass at accepting one or more data
files from the command line to be run against an awk program.

the `RuntimeConfig` has been updated to accept one or more (possibly
duplicate) file paths, which is set based on receiving one or more
positional arguments from the command line. any special handling of
mixing variable assignment with data file input has been saved for
another day (as rawk does not do the former at this time).

the runtime has been amended to iterate over the entirety of an awk
program for each piece of data (record) that has been received. this
allows for future implementations of `getline` to manipulate a reference
to a mutable iterator to retrieve additional records, without having to
rewrite a normal loop.

this commit does not handle the case of using the program file flag (-f)
with a data file. this functionality will come in a subesquent commit