feat(data): accept data files when -f provided (#18)
this commit extends the work done in 4e053d0 by allowing rawk to accept
data files to be read when the -f flag is present on the command line:
```bash
awk -f some_awk.awk some_data.dat
```

the problem that prevented this from being in the original featureset
that this commit works around is that the first positional argument must
be disambiguated. clap.rs has no idea whether the pos. arg is a file on
dist or a program literal, e.g.
```
// legal awk program, first pos. arg is a data file
awk -f some_awk.awk some_data.dat
// illegal awk program, first pos. arg is a program literal
awk -f some_awk.awk '{print $0;}'
```

to work around this (either the limitation of clap or more likely, my
lack of experience with it), if a first positional arg is provided in
addition to -f with a value, place that first positional arg in the
data_file_paths data structure, since its an illegal command.

this fix is a bit hacky - it assumes that this is the only logic for
positional arguments that will occur. it is very possible this will need
to be re-evaluated in the future as additional options are added.