feat(field_vars): print $0 if no expr provided
this commit adds support for the case where an action with a print
statement has a missing expression on the right hand side:
```awk
{ print; }
```

it does so by implicitly adding the 0th field variable (`$0`) to the
right hand side, expanding the previous statement like so:
```diff
- { print; }
+ { print $0; }
```

at this time, we still assume that a statement will end with a
semicolon, an assumption/bug that had existed in the implementation for
quite some time. in the future, we must check for an EoL or
non-expression (i.e. other keyword) to handle this more gracefully