chore(repo): remove actions-rs from actions runs (#47)
remove actions-rs from github actions runs for build_and_test, lint, and
format jobs. actions-rs has been deprecated/unmaintained since it was
archived of october 2023. while there are a few replacements that folks
are gravitating to, for this project (with its small size and toy-ish
nature), let's try running with the version of rust that gh action
runners have by default (which is relatively recent-ish).

 fix clippy errors from running against newer versions